refactor: 코드블록 프로젝트로 변경
