--- .gitignore
@@ -1,26 +1,0 @@
-# 빌드 디렉토리 무시
-/cmake-build-*/
-
-# IDE 관련 파일 무시
-/.idea
-*.iml
-
-# 컴파일된 객체 파일 및 실행 파일 무시
-*.o
-*.out
-*.a
-*.so
-*.exe
-
-# CMake 관련 파일 무시
-CMakeFiles/
-CMakeCache.txt
-cmake_install.cmake
-Makefile
-
-# 로그 및 임시 파일 무시
-*.log
-*.tmp
-
-# 코어 덤프 파일 무시
-core.*
