--- chess/event/PseudoCode.md
@@ -1,239 +1,0 @@
-## 1. 승급(Promotion)
-
-```
-(플레이어가 폰 기물을 이동한 상황)
-p <- 기물 구조체
-이동한 폰의 색상을 확인한다.
-IF 백색 플레이어인 경우
-    플레이어의 y축 좌표가 적팀 본진인지 확인한다
-    본진이 아니면 이벤트를 더 이상 진행하지 않는다 /* return 0 */
-    end
-ELSE IF 흑색 플레이어인 경우
-    플레이어의 y축 좌표가 적팀 본진인지 확인한다
-    본진이 아니면 이벤트를 더 이상 진행하지 않는다 /* return 0 */
-    end
-본진이면 다음 과정을 진행한다
-사용자에게 어떤 기물로 승급할 수 있는지 안내한다
-while(1)
-    newPiece <- 새 기물 구조체
-    사용자로부터 어떤 기물로 승급할 것인지 입력을 받는다
-    정상적으로 입력한 경우 반복문을 종료한다.
-    잘못된 입력일 경우 오류를 안내하고 재입력받는다
-    end
-새 기물 구조체를 생성한다(폰 기물을 치우고 동일한 위치에 생성)
-    새 기물 타입 = 사용자 입력값
-    새 기물 색상 = 기존 컬러와 동일
-    새 기물 현재 위치 = 기존 위치와 동일
-    그 외 데이터는 초기 값을 따른다.
-이벤트를 종료한다. /* return 1 */
-end
-```
-
----
-
-## 2. 앙파상(En Passant)
-
-```
-(플레이어가 폰 기물을 선택한 상황)
-p <- 선택한 폰 기물 구조체
-board <- 체스판
-turn <- 현재 진행하고 있는 턴 수
-선택한 폰의 색상과 위치를 확인한다
-IF 백색 플레이어이고 현재 y축이 3인 경우
-OR 흑색 플레이어이고 현재 y축이 4인 경우
-    True : 다음 알고리즘 진행
-    False : 이벤트 종료 /* return 0 */
-    end
-기물의 양쪽에 적 폰이 존재하는 지 확인한다
-left <- 기물의 왼쪽 칸
-right <- 기물의 오른쪽 칸
-IF 왼쪽칸에 적 폰 기물이 존재하는 경우
-OR 오른쪽칸에 적 폰 기물이 존재하는 경우
-    True : 다음 알고리즘 실행
-    opponent <- 적 기물 구조체
-    적 폰 기물이 딱 한 번만 이동했는지, 그리고 방금 움직인 건지 확인
-    IF 적 폰 기물이 딱 한 번만 이동했는가?
-    AND 적 폰 기물이 직전 턴에 움직였는가?
-        True : 앙파상 기준 충족
-        사용자 입력 함수 실행 후 알고리즘 종료 /* return 1 */
-        False : 이벤트 종료  
-        end
-    False : 이벤트 종료
-    end
-end
-```
-
----
-
-## 3. 캐슬링(Castling)
-
-```
-(플레이어가 킹 기물을 선택한 상황)
-p <- 선택한 킹 기물 구조체
-board <- 체스판
-선택한 킹의 이동 전적을 확인
-IF 킹이 한 번이라도 움직였는가?
-    TRUE : 이벤트 종료
-    END
-킹이 안전한 지 확인
-IF 현재 킹이 '체크' 상태인가?
-    TRUE : 이벤트 종료
-    END
-킹의 양쪽 방향을 1칸씩 점검한다.
-점검하려는 칸에 기물이 존재하는지 확인한다.
-IF 기물이 존재하는가?
-    FALSE : 다음 칸 점검
-    TRUE : 다음 조건 확인
-    IF 해당 기물이 '룩'이 아닌 다른 기물인가?
-        TRUE : 이벤트 종료
-        END
-    해당 방향의 룩 기물이 한 번이라도 움직였는지 확인
-    ELSE IF '룩' 기물이 한 번이라도 움직였는가?
-        TRUE : 이벤트 종료
-        END
-    해당 방향을 이동 가능 방향으로 추가
-    END
-사용자에게 캐슬링이 가능함을 알린다.
-사용자에게 캐슬링 진행 가능한 방향을 알려주고, 입력을 받는다.
-사용자는 입력을 통해 캐슬링 진행 또는 거부를 결정한다.
-SWITCH 사용자 선택
-    CASE LEFT : 킹은 왼쪽으로 2칸 이동, 룩은 킹의 바로 오른쪽 칸에 배치
-    이벤트 완료 및 캐슬링 진행을 알림
-    CASE RIGHT : 킹은 오른쪽으로 2칸 이동, 룩은 킹의 바로 왼쪽 칸에 배치
-    이벤트 완료 및 캐슬링 진행을 알림
-    CASE CANCEL : 이벤트 종료
-    END
-END
-```
-
----
-
-## 4. 기물별 이동 경로 계산(calculate_moves)
-
-```
-(매 턴마다 사용. 상대 기물의 공격 경로를 확인하여 게임 진행 여부를 판단하기 위함)
-p <- 경로 확인 대상 기물
-board <- 체스판
-opponent_possible_move <- 상대의 공격 가능 경로를 모을 배열
-현재 확인 중인 기물의 타입을 확인한다.
-SWITCH : 기물 타입
-    CASE PAWN :
-        양측 대각선 전진 경로 확인
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    CASE KNIGHT :
-        나이트의 이동 경로 확인
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    CASE ROOK :
-        상하좌우 이동 경로 확인
-        각 방향마다 1칸씩 이동하며 기물이 존재하거나 체스판을 벗어날 때까지 점검
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    CASE BISHOP :
-        대각선 이동 경로 확인
-        각 방향마다 1칸씩 이동하며 기물이 존재하거나 체스판을 벗어날 때까지 점검
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    CASE QUEEN :
-        8방향 이동 경로 확인
-        각 방향마다 1칸씩 이동하며 기물이 존재하거나 체스판을 벗어날 때까지 점검
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    CASE KING :
-        8방향 이동 경로 확인
-        IF 해당 좌표가 비어있는가?
-        OR 아직 확인하지 않은 좌표인가?
-            TRUE : 공격 가능 경로에 추가
-            END
-        이벤트 종료
-    END
-END
-```
-
----
-
-## 5. 킹의 안전 여부 확인(is_king_safe)
-
-```
-king <- 플레이어의 킹 기물
-board <- 체스판
-opponent_possible_move <- 적 공격 가능 경로 모음
-킹 기물의 좌표 확인
-현재 좌표가 적의 위협을 받고 있는지 확인.
-IF 킹이 상대의 공격 가능 경로 내에 있는가?
-    TRUE : 1
-    FALSE : 0
-    END
-END
-```
-
----
-
-## 6. 체크메이트 판단 함수(validate_checkmate)
-
-```
-(킹이 안전하지 않은 경우 해당 함수를 사용한다.)
-board <- 체스판
-opponent <- 상대 픓레이어 색상
-opponent_possible_move <- 상대 공격 가능 경로 모음
-킹이 이동할 수 있는 8칸 중, 안전한 위치가 있는지 확인한다.
-IF 해당 좌표가 비어있는가?
-AND 해당 좌표가 상대 공격 경로 밖에 있는가?
-    TRUE : "Check" 반환
-    END
-킹의 이동이 불가능한 경우 다음 알고리즘을 진행한다.
-각 아군 기물들의 이동 시뮬레이션을 체크하여 아군 기물이 킹을 방어할 수 있는지 확인한다.
-IF 아군 기물이 상대의 공격을 방어할 수 있는가?
-    TRUE : "Check" 반환
-    END
-더 이상 상대의 공격으로부터 킹을 방어할 수단이 없음을 확인하였다.
-"CheckMate" 반환
-END
-```
-
----
-
-## 7. 스테일메이트 판단 함수(validate_stalemate)
-
-```
-(킹이 안전한 경우 해당 함수를 사용한다.)
-board <- 체스판
-opponent <- 상대 플레이어 색상
-opponenet_possible_move <- 상대 공격 가능 경로 모음
-king_arounded <- 킹이 완전히 아군에게 둘러싸여 있는지 체크
-킹의 8방향 이동 경로가 한 곳이라도 안전한 지 확인한다.
-IF 해당 좌표가 비어있는가?
-    킹이 둘러싸여있지 않으므로 king_arounded는 false
-    IF 해당 좌표가 상대 공격 경로 밖에 있는가?
-        TRUE : "NONE" 반환
-        END
-    END
-IF 킹이 완전히 아군에게 보호받고 있는가?
-    TRUE : "NONE" 반환
-    END
-킹이 움직이면 반드시 상대에게 공격당하는 상황임이 증명된다.
-다른 아군 기물이 이동할 수 있는지 확인한다.
-각 아군 기물의 이동 경로를 확인한다.
-IF 해당 기물이 움직일 공간이 있는가?
-    TRUE : "NONE" 반환
-    END
-더 이상 움직일 있는 기물이 존재하지 않는다.
-"StaleMate" 반환
-END
-```
