--- chess/readme.md
@@ -1,226 +1,0 @@
-> **pesudo code**: 폰 이동 검증 함수
-> ```c
-> 
->   1. 현재 위치의 말 정보를 가져옵니다(pieces 의 인덱스를 가져와서 기물 검증)
->   - 폰의 색상(백 'W' 또는 흑 'B') 을 확인힙니다.
->
->   2. 백 폰의 두 칸 이동 검증:
->    - 현재 위치가 6번째 행인지 확인합니다
->    - 이동할 위치가 두 칸 위로 이동한 위치인지 확인합니다
->    - 만약 경로가 비어 있다면 이동을 허용하고 경로가 비어있지 않다면 0을 반환합니다.
->
->   3. 흑 폰의 두 칸 이동 검증:
->    - 현재 위치가 1번째 행인지 확인합니다
->    - 이동할 위치가 두 칸 아래로 이동한 위치인지 확인합니다
->    - 만약 경로가 비어 있다면 이동을 허용하고 경로가 비어있지 않다면 0을 반환합니다.
->
->   4. 백 폰의 한 칸 이동 검증:
->    - 이동할 위치가 현재 위치의 행보다 1이 작은지 확인합니다
->    - 만약 경로가 비어 있다면 이동을 허용하고 경로가 비어있지 않다면 0을 반환합니다.
->
->   5. 흑 폰의 한 칸 이동 검증:
->    - 목표 위치가 한 칸 아래로 이동한 위치인지 확인합니다
->    - 만약 경로가 비어 있다면 이동을 허용하고 경로가 비어있지 않다면 0을 반환합니다.
->
->   6. 대각선 공격 이동 검증:
->    - 백 폰의 경우:
->        - 목표 위치가 왼쪽 위나 오른쪽 위 대각선에 있는지 확인합니다
->        - 이동할 위치에 상대 기물이 있다면 이동을 허용합니다
-> 
->    - 흑 폰의 경우:
->        - 목표 위치가 왼쪽 아래나 오른쪽 아래 대각선에 있는지 확인합니다
->        - 목표 위치에 상대 기물이 있다면 이동을 허용합니다
->
->   7. 위 조건을 모두 통과하지 못하면 이동및 공격을 허용하지 않습니다.
-
----------------------------------------
-
----------------------------------------
-
-> **pesudo code**: 폰 이동 함수 
-> ```c
->   1. 폰의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. pawn의 인덱스가 '\0' 이 아니고 기물이 폰일때 if문을 수행
->   - 목표 위치에 현재 폰의 위치를 저장.
->   - 현재 현재 폰의 위치를 '.' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 폰이 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
-
--------------------------------------
-
--------------------------------------
-
-> **pesudo code**: 나이트 이동 검증 함수
-> ```c
-> is_valid_knight_move(Position from, Position to, ChessBoard *board) 함수
->   방향 배열 입력: {{1, 2}, {2, 1}, {2, -1}, {1, -2}, {-1, 2}, {-2, 1}, {-2, -1}, {-1, -2}}
->   for문 실행 각 방향에 대해 검증
->       newX: X의 새로운 위치 계산
->       newY: Y의 새로운 위치 계산
->       is_within_board 함수를 사용하여 newX, newY 가 체스판 밖이라면 continue
->       newX, newY 가 목표 위치일때 목표 위치에 기물이 없거나 상대 기물(색이 다른 기물이 있다면)
->       1반환
-> 반복 종료 후 0 반환
-> 
-
------------------------------------------
-
------------------------------------------
-
-> **pesudo code**: 나이트 이동 함수
-> ```c
->   1. 나이트의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. 나이트의 인덱스가 '\0' 이 아니고 기물이 폰일때 if문을 수행
->   - 목표 위치에 현재 나이트의 위치를 저장합니다
->   - 현재 현재 나이트의 위치를 '\0' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 나이트가 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
-
-----------------------------------------
-
-----------------------------------------
-
-> **pesudo code**: 룩 이동 검증 함수
-> ```c
-> is_valid_rook_move(Position from, Position to, ChessBoard *board)시작 위치, 목표 위치, 체스 구조체
->   이동 방향 배열 입력: {{1, 0}, {-1, 0}, {0, 1}, {0, -1}}
->   for 반복문 실행: 각 진행 방향에 따른 이동 가능에 대한 검증(0부터 3까지 반복하여 4가지 방향 전체 검증)
->       x에 최초 x값 , y의 최초 y값 초기화 후 
->       이동 방향을 결정하는 newX,newY를 각각 초기화 합니다(배열의 각 열의 인덱스를 입력)
->       while 반복문 (is_within_board(x, y) 함수를 이용해서 체스 판 내부에 있을때까지만 반복)
->           x,y 에 각각 newX,newY 씩 더해서 1칸씩 이동
->           경로에 기물이 있는 경우 break를 합니다
->       이때 목표한 위치에 다다른 경우
->       from_piece_index에 기물 찾기 함수를 이용하여 찾은 기물 인덱스를 초기화 하고
->       target_piece_index에 기물 찾기 함수(못찾을시 0 반환)를 이용하여 찾은 기물 인덱스를 초기화 합니다. 
->       조건문을 사용해서 인덱스에 기물이 없거나 상대 기물이 있는 경우 1을 반환합니다
-> 전체 조건에서 맞는 조건이 없는 경우, 0을 반환합니다
-
-
----------------------------------------
-
----------------------------------------
-
-> **pesudo code**: 룩 이동 함수
-> ```c
->   1. 룩의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. rook의 인덱스가 '\0' 이 아니고 기물이 룩일때 if문을 수행
->   - 목표 위치에 현재 룩의 위치를 저장합니다
->   - 현재 현재 룩의 위치를 '\0' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 룩이 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
-
----------------------------------------
-
----------------------------------------
-
-> **pesudo code**: 비숍 이동 검증 함수
-> ```c
-> is_valid_bishop_move(Position from, Position to, ChessBoard *board)시작 위치, 목표 위치, 체스 구조체
-> 이동 방향 배열 입력: {{1,1}, {-1,-1}, {1,-1}, {-1,1}}
->   for 반복문 실행: 각 진행 방향에 따른 이동 가능에 대한 검증(0부터 3까지 반복하여 4가지 방향 전체 검증)
->       x에 최초 x값 , y의 최초 y값 초기화 후 
->       이동 방향을 결정하는 newX,newY를 각각 초기화 합니다(배열의 각 열의 인덱스를 입력)
->       while 반복문 (is_within_board(x, y) 함수를 이용해서 체스 판 내부에 있을때까지만 반복)
->           x,y 에 각각 newX,newY 씩 더해서 1칸씩 이동
->           경로에 기물이 있는 경우 break를 합니다
->       이때 목표한 위치에 다다른 경우
->       from_piece_index에 기물 찾기 함수를 이용하여 찾은 기물 인덱스를 초기화 하고
->       target_piece_index에 기물 찾기 함수(못찾을시 0 반환)를 이용하여 찾은 기물 인덱스를 초기화 합니다. 
->       조건문을 사용해서 인덱스에 기물이 없거나 상대 기물이 있는 경우 1을 반환합니다
-
--------------------------------------
-
--------------------------------------
-
-> **pesudo code**: 비숍 이동 함수
-> ```c
->   1. 비숍의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. 비숍의 인덱스가 '\0' 이 아니고 기물이 룩일때 if문을 수행
->   - 목표 위치에 현재 비숍의 위치를 저장합니다
->   - 현재 현재 비숍의 위치를 '\0' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 비숍이 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
-
---------------------------------------
-
---------------------------------------
-
-> **pesudo code**: 퀸 이동 검증 함수
-> ```c
-> is_valid_queen_move(Position from, Position to, ChessBoard *board)시작 위치, 목표 위치, 체스 구조체
-> 퀸이 룩과 비숍의 이동 방식을 합친 기물이기 때문에
-> 룩 이동 검증 함수와 비숍 이동 검증 함수를 동시에 실행하여서 OR연산자를 이용해 두 함수 중 하나가 가능하면 1을 반환합니다
-> 두 함수 모두 0을 반환했을 때 0을 반환합니다
-
---------------------------------------
-
---------------------------------------
-
-> **pesudo code**: 퀸 이동 함수
-> ```c
->   1. 퀸의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. 퀸의 인덱스가 '\0' 이 아니고 기물이 퀸일때 if문을 수행
->   - 목표 위치에 현재 나이트의 위치를 저장합니다
->   - 현재 현재 나이트의 위치를 '\0' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 퀸이 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
-
-------------------------------------
-
-------------------------------------
-
-> **pesudo code**: 킹 이동 검증 함수(임시)
-> ```c
-> is_valid_king_move(Position from, Position to, ChessBoard *board) 함수
->   방향 배열 입력: {{0, 1}, {0, -1}, {-1, 0}, {1, 0}, {-1, 1}, {1, 1}, {-1, -1}, {1, -1}}
->   for문 실행 각 방향에 대해 검증
->       newX: X의 새로운 위치 계산
->       newY: Y의 새로운 위치 계산
->       is_within_board 함수를 사용하여 newX, newY 가 체스판 밖이라면 continue
->       newX, newY 가 목표 위치일때 목표 위치에 기물이 없거나 상대 기물(색이 다른 기물이 있다면)
->   1반환
->   반복 종료 후 0 반환
-
---------------------------------------
-
---------------------------------------
-
-> **pesudo code**: 킹 이동 함수(임시)
-> ```c
->   1. 킹의 인덱스를 가져올 find_piece_at_position 함수 사용
->   - x y 좌표 및 board 구조체 인자에 입력
->   
->   2. 킹의 인덱스가 '\0' 이 아니고 기물이 퀸일때 if문을 수행
->   - 목표 위치에 현재 나이트의 위치를 저장합니다
->   - 현재 현재 나이트의 위치를 '\0' 으로 바꿔줍니다.
->   - pieces구조체 변수의 위치 정보를 바꿔줍니다.
-> 
->   3. 킹이 어디서 어디로 이동했는지 출력해줍니다.
->   
->   4. 턴 진행 수를 증감식으로 추가합니다.
