--- jang han/chess/슈도코드(1).txt
@@ -1,29 +1,0 @@
-
-1. 현재 상태 표시
-   - 체스판 출력
-   - 현재 턴과 플레이어 표시
-3. 플레이어 턴 처리
-   - turn이 홀수면 player1(board) 호출  
-   - 짝수면 player2(board) 호출  
-
-4. 특수 이동 처리
-   - 승급(Promotion)
-     - 각 폰을 검사하여 상대 진영 끝에 도달했는지 확인  
-     - 승급 조건 만족 시 promotion(board, pawn) 호출  
-   - 앙파상(En Passant)
-     - 각 폰을 검사하여 앙파상 조건 확인  
-     - 조건 만족 시 en_passant(board, pawn) 호출  
-   - 캐슬링(Castling)
-     - 현재 킹의 캐슬링 가능 여부를 확인  
-     - 조건 만족 시 castling(board, king, is_checked) 호출  
-
-5. 킹 생존 확인
-   - 백 킹과 흑 킹이 모두 살아 있는지 확인  
-   - 킹이 잡힌 경우 상대 팀 승리 메시지 출력 및 게임 종료  
-
-6. 사용자 입력
-   - 사용자로부터 종료(`q`) 또는 재시작(`r`) 선택 입력
-   - q: 승리 메시지 출력 후 게임 종료  
-   - r: 체스판 초기화 및 턴 수 초기화 후 게임 재시작  
-   턴 증가
-   - 턴 수 증가
