+++ pseudo_event.txt
@@ -1,0 +1,28 @@
+< 1. 승급(Promotion) >
+
+(플레이어가 폰 기물을 이동한 상황)
+p <- 기물 구조체
+이동한 폰의 색상을 확인한다.
+IF 백색 플레이어인 경우
+    플레이어의 y축 좌표가 적팀 본진인지 확인한다
+    본진이 아니면 이벤트를 더 이상 진행하지 않는다 /* return 0 */
+    end
+ELSE IF 흑색 플레이어인 경우
+    플레이어의 y축 좌표가 적팀 본진인지 확인한다
+    본진이 아니면 이벤트를 더 이상 진행하지 않는다 /* return 0 */
+    end
+본진이면 다음 과정을 진행한다
+사용자에게 어떤 기물로 승급할 수 있는지 안내한다
+while(1)
+    newPiece <- 새 기물 구조체
+    사용자로부터 어떤 기물로 승급할 것인지 입력을 받는다
+    정상적으로 입력한 경우 반복문을 종료한다.
+    잘못된 입력일 경우 오류를 안내하고 재입력받는다
+    end
+새 기물 구조체를 생성한다(폰 기물을 치우고 동일한 위치에 생성)
+    새 기물 타입 = 사용자 입력값
+    새 기물 색상 = 기존 컬러와 동일
+    새 기물 현재 위치 = 기존 위치와 동일
+    그 외 데이터는 초기 값을 따른다.
+이벤트를 종료한다. /* return 1 */
+end
