--- second code
@@ -1,22 +1,0 @@
-1. main()
-프로그램의 시작 지점입
-게임의 환영 메시지를 출력하고, 화면을 지운 후 게임 루프를 시작
-x 변수를 사용하여 턴을 관리하며, 홀수 턴은 플레이어 1, 짝수 턴은 플레이어 2의 차례로 진행
-각 턴 후 화면을 지우고 체스판을 표시
-2. display()
-현재 체스판의 상태를 출력하는 함수
-8x8 체스판을 나타내며, 각 행과 열의 인덱스를 표시
-체스말과 빈 공간을 '||'로 구분하여 출력
-3. change(int r1, int c1, int r2, int c2)
-체스판에서 두 위치의 말을 교환하는 함수
-r1, c1은 원래 위치, r2, c2는 이동할 위치
-두 위치의 값을 서로 바꾼다
-4. player1()
-플레이어 1의 턴을 처리하는 함수
-사용자가 이동할 말을 선택하고, 해당 말의 종류에 따라 각각의 말에 대한 함수를 호출
-사용자가 이동할 목표 위치를 입력받고, change 함수를 통해 말을 이동시킵니다.
-잘못된 위치를 입력할 경우 다시 입력을 받도록 합니다.
-5. player2()
-플레이어 2의 턴을 처리하는 함수
-플레이어 1과 유사하게 작동하지만 대문자를 사용하는 말에 대해 처리
-잘못된 위치를 입력할 경우 다시 입력을 받도록 함
